requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
ffmpeg-python==0.2.0
whisper.cpp-cli
openai
python-dotenv
